lib cf-socket
win32:
```
 177 |     vals.keepalivetime = optval;
      |                        ~ ^~~~~~
/home/runner/work/curl/curl/curl/lib/cf-socket.c:180:30: error: implicit conversion changes signedness: 'int' to 'u_long' (aka 'unsigned long') [-Werror,-Wsign-conversion]
  180 |     vals.keepaliveinterval = optval;
      |                            ~ ^~~~~~
```
Ref: https://github.com/curl/curl/actions/runs/8862225447/job/24334941063?pr=13489#step:3:6125

+++ lib/cf-socket.c
@@ -1450,7 +1450,7 @@ static ssize_t cf_socket_send(struct Curl_cfilter *cf, struct Curl_easy *data,
     }
   }
   if(cf->cft != &Curl_cft_udp && ctx->wpartial_percent > 0 && len > 8) {
-    len = len * ctx->wpartial_percent / 100;
+    len = len * (size_t)ctx->wpartial_percent / 100;
     if(!len)
       len = 1;
     CURL_TRC_CF(data, cf, "send(len=%zu) SIMULATE partial write of %zu bytes",
@@ -2051,7 +2051,7 @@ static void set_accepted_remote_ip(struct Curl_cfilter *cf,
   ctx->ip.remote_ip[0] = 0;
   ctx->ip.remote_port = 0;
   plen = sizeof(ssrem);
-  memset(&ssrem, 0, plen);
+  memset(&ssrem, 0, (size_t)plen);
   if(getpeername(ctx->sock, (struct sockaddr*) &ssrem, &plen)) {
     int error = SOCKERRNO;
     failf(data, "getpeername() failed with errno %d: %s",
